Simplify grizzly instrumentation
