+++ instrumentation/grizzly-2.0/grizzly-2.0.gradle
@@ -21,6 +21,8 @@ testSets {
 dependencies {
   compileOnly group: 'org.glassfish.grizzly', name: 'grizzly-http-server', version: '2.0'
 
+  compile project(':instrumentation:grizzly-http-2.0')
+
   testCompile group: 'javax.xml.bind', name: 'jaxb-api', version: '2.2.3'
   testCompile group: 'javax.ws.rs', name: 'javax.ws.rs-api', version: '2.0'
   testCompile group: 'org.glassfish.grizzly', name: 'grizzly-http-server', version: '2.0'
