+++ instrumentation/grizzly-2.0/src/test/groovy/GrizzlyTest.groovy
@@ -14,16 +14,18 @@
  * limitations under the License.
  */
 
+
 import io.opentelemetry.auto.test.base.HttpServerTest
+import org.glassfish.grizzly.http.server.HttpServer
+import org.glassfish.jersey.grizzly2.httpserver.GrizzlyHttpServerFactory
+import org.glassfish.jersey.server.ResourceConfig
+
 import javax.ws.rs.GET
 import javax.ws.rs.NotFoundException
 import javax.ws.rs.Path
 import javax.ws.rs.QueryParam
 import javax.ws.rs.core.Response
 import javax.ws.rs.ext.ExceptionMapper
-import org.glassfish.grizzly.http.server.HttpServer
-import org.glassfish.jersey.grizzly2.httpserver.GrizzlyHttpServerFactory
-import org.glassfish.jersey.server.ResourceConfig
 
 import static io.opentelemetry.auto.test.base.HttpServerTest.ServerEndpoint.ERROR
 import static io.opentelemetry.auto.test.base.HttpServerTest.ServerEndpoint.EXCEPTION
@@ -104,10 +106,5 @@ class GrizzlyTest extends HttpServerTest<HttpServer> {
       }
       return null
     }
-  }
-
-  @Override
-  String expectedOperationName(String method, ServerEndpoint serverEndpoint) {
-    return 'HttpHandler.doHandle'
   }
 }
