+++ instrumentation/grizzly-http-2.0/src/main/java/io/opentelemetry/auto/instrumentation/grizzly/DefaultFilterChainAdvice.java
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.auto.instrumentation.grizzly.http.v2_3;
+package io.opentelemetry.auto.instrumentation.grizzly;
 
 import net.bytebuddy.asm.Advice;
 import org.glassfish.grizzly.filterchain.FilterChainContext;
-import org.glassfish.grizzly.http.HttpResponsePacket;
 
-public class HttpServerFilterAdvice {
-  @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-  public static void onExit(
+public class DefaultFilterChainAdvice {
+
+  @Advice.OnMethodEnter(suppress = Throwable.class)
+  public static void onFail(
       @Advice.Argument(0) final FilterChainContext ctx,
-      @Advice.Argument(2) final HttpResponsePacket response) {
-    GrizzlyDecorator.onHttpServerFilterPrepareResponseExit(ctx, response);
+      @Advice.Argument(1) final Throwable throwable) {
+    GrizzlyDecorator.onFilterChainFail(ctx, throwable);
   }
 }
