+++ instrumentation/grizzly-http-2.0/src/main/java/io/opentelemetry/auto/instrumentation/grizzly/FilterInstrumentation.java
@@ -14,35 +14,48 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.auto.instrumentation.grizzly.http.v2_3;
+package io.opentelemetry.auto.instrumentation.grizzly;
 
-import static net.bytebuddy.matcher.ElementMatchers.isPrivate;
+import static io.opentelemetry.auto.tooling.ClassLoaderMatcher.hasClassesNamed;
+import static java.util.Collections.singletonMap;
+import static net.bytebuddy.matcher.ElementMatchers.hasSuperClass;
+import static net.bytebuddy.matcher.ElementMatchers.isPublic;
 import static net.bytebuddy.matcher.ElementMatchers.named;
+import static net.bytebuddy.matcher.ElementMatchers.not;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 
 import com.google.auto.service.AutoService;
 import io.opentelemetry.auto.tooling.Instrumenter;
-import java.util.Collections;
 import java.util.Map;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
+import net.bytebuddy.matcher.ElementMatchers;
 
 @AutoService(Instrumenter.class)
-public class HttpServerFilterInstrumentation extends Instrumenter.Default {
+public final class FilterInstrumentation extends Instrumenter.Default {
 
-  public HttpServerFilterInstrumentation() {
-    super("grizzly-filterchain");
+  public FilterInstrumentation() {
+    super("grizzly");
+  }
+
+  @Override
+  public ElementMatcher<ClassLoader> classLoaderMatcher() {
+    return hasClassesNamed("org.glassfish.grizzly.filterchain.BaseFilter");
   }
 
   @Override
   public ElementMatcher<? super TypeDescription> typeMatcher() {
-    return named("org.glassfish.grizzly.http.HttpServerFilter");
-  }
-
-  @Override
-  protected boolean defaultEnabled() {
-    return false;
+    return hasSuperClass(named("org.glassfish.grizzly.filterchain.BaseFilter"))
+        // HttpCodecFilter is instrumented in the server instrumentation
+        .and(
+            not(
+                ElementMatchers.<TypeDescription>named(
+                    "org.glassfish.grizzly.http.HttpCodecFilter")))
+        .and(
+            not(
+                ElementMatchers.<TypeDescription>named(
+                    "org.glassfish.grizzly.http.HttpServerFilter")));
   }
 
   @Override
@@ -51,14 +64,16 @@ public class HttpServerFilterInstrumentation extends Instrumenter.Default {
   }
 
   @Override
+  protected boolean defaultEnabled() {
+    return false;
+  }
+
+  @Override
   public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-    return Collections.singletonMap(
-        named("prepareResponse")
+    return singletonMap(
+        named("handleRead")
             .and(takesArgument(0, named("org.glassfish.grizzly.filterchain.FilterChainContext")))
-            .and(takesArgument(1, named("org.glassfish.grizzly.http.HttpRequestPacket")))
-            .and(takesArgument(2, named("org.glassfish.grizzly.http.HttpResponsePacket")))
-            .and(takesArgument(3, named("org.glassfish.grizzly.http.HttpContent")))
-            .and(isPrivate()),
-        packageName + ".HttpServerFilterAdvice");
+            .and(isPublic()),
+        packageName + ".FilterAdvice");
   }
 }
