+++ instrumentation/grizzly-http-2.0/src/main/java/io/opentelemetry/auto/instrumentation/grizzly/HttpCodecFilterAdvice.java
@@ -14,17 +14,18 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.auto.instrumentation.grizzly.http.v2_3;
+package io.opentelemetry.auto.instrumentation.grizzly;
 
 import net.bytebuddy.asm.Advice;
 import org.glassfish.grizzly.filterchain.FilterChainContext;
+import org.glassfish.grizzly.http.HttpHeader;
 
-public class DefaultFilterChainAdvice {
+public class HttpCodecFilterAdvice {
 
-  @Advice.OnMethodEnter(suppress = Throwable.class)
-  public static void onFail(
+  @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
+  public static void onExit(
       @Advice.Argument(0) final FilterChainContext ctx,
-      @Advice.Argument(1) final Throwable throwable) {
-    GrizzlyDecorator.onFilterChainFail(ctx, throwable);
+      @Advice.Argument(1) final HttpHeader httpHeader) {
+    GrizzlyDecorator.onHttpCodecFilterExit(ctx, httpHeader);
   }
 }
