+++ instrumentation/grizzly-http-2.0/src/main/java/io/opentelemetry/auto/instrumentation/grizzly/HttpCodecFilterInstrumentation.java
@@ -14,48 +14,35 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.auto.instrumentation.grizzly.http.v2_3;
+package io.opentelemetry.auto.instrumentation.grizzly;
 
-import static io.opentelemetry.auto.tooling.ClassLoaderMatcher.hasClassesNamed;
-import static java.util.Collections.singletonMap;
-import static net.bytebuddy.matcher.ElementMatchers.hasSuperClass;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
 import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.not;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 
 import com.google.auto.service.AutoService;
 import io.opentelemetry.auto.tooling.Instrumenter;
+import java.util.HashMap;
 import java.util.Map;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
-import net.bytebuddy.matcher.ElementMatchers;
 
 @AutoService(Instrumenter.class)
-public final class FilterInstrumentation extends Instrumenter.Default {
+public final class HttpCodecFilterInstrumentation extends Instrumenter.Default {
 
-  public FilterInstrumentation() {
-    super("grizzly-filterchain");
-  }
-
-  @Override
-  public ElementMatcher<ClassLoader> classLoaderMatcher() {
-    return hasClassesNamed("org.glassfish.grizzly.filterchain.BaseFilter");
+  public HttpCodecFilterInstrumentation() {
+    super("grizzly");
   }
 
   @Override
   public ElementMatcher<? super TypeDescription> typeMatcher() {
-    return hasSuperClass(named("org.glassfish.grizzly.filterchain.BaseFilter"))
-        // HttpCodecFilter is instrumented in the server instrumentation
-        .and(
-            not(
-                ElementMatchers.<TypeDescription>named(
-                    "org.glassfish.grizzly.http.HttpCodecFilter")))
-        .and(
-            not(
-                ElementMatchers.<TypeDescription>named(
-                    "org.glassfish.grizzly.http.HttpServerFilter")));
+    return named("org.glassfish.grizzly.http.HttpCodecFilter");
+  }
+
+  @Override
+  protected boolean defaultEnabled() {
+    return false;
   }
 
   @Override
@@ -65,10 +52,21 @@ public final class FilterInstrumentation extends Instrumenter.Default {
 
   @Override
   public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-    return singletonMap(
+    final Map<ElementMatcher<? super MethodDescription>, String> transformers = new HashMap<>();
+    // this is for 2.3 through 2.3.19
+    transformers.put(
         named("handleRead")
             .and(takesArgument(0, named("org.glassfish.grizzly.filterchain.FilterChainContext")))
+            .and(takesArgument(1, named("org.glassfish.grizzly.http.HttpPacketParsing")))
             .and(isPublic()),
-        packageName + ".FilterAdvice");
+        packageName + ".HttpCodecFilterOldAdvice");
+    // this is for 2.3.20+
+    transformers.put(
+        named("handleRead")
+            .and(takesArgument(0, named("org.glassfish.grizzly.filterchain.FilterChainContext")))
+            .and(takesArgument(1, named("org.glassfish.grizzly.http.HttpHeader")))
+            .and(isPublic()),
+        packageName + ".HttpCodecFilterAdvice");
+    return transformers;
   }
 }
