+++ instrumentation/grizzly-http-2.0/src/main/java/io/opentelemetry/auto/instrumentation/grizzly/HttpCodecFilterOldAdvice.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.auto.instrumentation.grizzly.http.v2_3;
+package io.opentelemetry.auto.instrumentation.grizzly;
 
 import net.bytebuddy.asm.Advice;
 import org.glassfish.grizzly.filterchain.FilterChainContext;
