+++ instrumentation/grizzly-http-2.0/src/test/groovy/GrizzlyFilterchainServerTestInstrumentation.java
@@ -15,6 +15,7 @@
  */
 
 import static net.bytebuddy.matcher.ElementMatchers.named;
+import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 
 import com.google.auto.service.AutoService;
 import io.opentelemetry.auto.test.base.HttpServerTestAdvice;
@@ -22,14 +23,19 @@ import io.opentelemetry.auto.tooling.Instrumenter;
 import net.bytebuddy.agent.builder.AgentBuilder;
 
 @AutoService(Instrumenter.class)
-public class GrizzlyTestInstrumentation implements Instrumenter {
+public class GrizzlyFilterchainServerTestInstrumentation implements Instrumenter {
 
   @Override
   public AgentBuilder instrument(final AgentBuilder agentBuilder) {
     return agentBuilder
-        .type(named("org.glassfish.grizzly.http.server.HttpHandlerChain"))
+        .type(named("org.glassfish.grizzly.http.HttpCodecFilter"))
         .transform(
             new AgentBuilder.Transformer.ForAdvice()
-                .advice(named("doHandle"), HttpServerTestAdvice.ServerEntryAdvice.class.getName()));
+                .advice(
+                    named("handleRead")
+                        .and(
+                            takesArgument(
+                                0, named("org.glassfish.grizzly.filterchain.FilterChainContext"))),
+                    HttpServerTestAdvice.ServerEntryAdvice.class.getName()));
   }
 }
