+++ instrumentation/grizzly-http-2.0/src/test/groovy/GrizzlyTest.groovy
@@ -35,7 +35,7 @@ import static io.opentelemetry.auto.test.base.HttpServerTest.ServerEndpoint.SUCC
 class GrizzlyTest extends HttpServerTest<HttpServer> {
 
   static {
-    System.setProperty("ota.integration.grizzly-filterchain.enabled", "true")
+    System.setProperty("ota.integration.grizzly.enabled", "true")
   }
 
   @Override
