--- instrumentation/grizzly-http-2.3/src/main/java/io/opentelemetry/auto/instrumentation/grizzly/http/v2_3/ExtractAdapter.java
@@ -1,29 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.opentelemetry.auto.instrumentation.grizzly.http.v2_3;
-
-import io.opentelemetry.context.propagation.HttpTextFormat;
-import org.glassfish.grizzly.http.HttpHeader;
-
-public class ExtractAdapter implements HttpTextFormat.Getter<HttpHeader> {
-  public static final ExtractAdapter GETTER = new ExtractAdapter();
-
-  @Override
-  public String get(final HttpHeader carrier, final String key) {
-    return carrier.getHeader(key);
-  }
-}
