+++ settings.gradle
@@ -66,7 +66,7 @@ include ':instrumentation:geode-1.4'
 include ':instrumentation:google-http-client-1.19'
 include ':instrumentation:grizzly-2.0'
 include ':instrumentation:grizzly-client-1.9'
-include ':instrumentation:grizzly-http-2.3'
+include ':instrumentation:grizzly-http-2.0'
 include ':instrumentation:grpc-1.5'
 include ':instrumentation:hibernate:hibernate-3.3'
 include ':instrumentation:hibernate:hibernate-4.0'
